backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
Faker==35.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.0
